Updated Cordova for IA SDK version 2.3
