--- src/ios/IndoorAtlas2.2/IndoorAtlas.framework/Headers
@@ -1,1 +1,0 @@
-Versions/Current/Headers
