--- src/ios/IndoorAtlas2.2/IndoorAtlas.framework/IndoorAtlas
@@ -1,1 +1,0 @@
-Versions/Current/IndoorAtlas
