--- src/ios/IndoorAtlas2.2/IndoorAtlas.framework/Resources
@@ -1,1 +1,0 @@
-Versions/Current/Resources
