--- src/ios/IndoorAtlas2.2/IndoorAtlas.framework/Versions/Current
@@ -1,1 +1,0 @@
-2.2.3-32
