+++ src/ios/IndoorAtlas2.3/IndoorAtlas.framework/Headers
@@ -1,0 +1,1 @@
+Versions/Current/Headers
