+++ src/ios/IndoorAtlas2.3/IndoorAtlas.framework/IndoorAtlas
@@ -1,0 +1,1 @@
+Versions/Current/IndoorAtlas
