+++ src/ios/IndoorAtlas2.3/IndoorAtlas.framework/Modules
@@ -1,0 +1,1 @@
+Versions/Current/Modules
