+++ src/ios/IndoorAtlas2.3/IndoorAtlas.framework/Resources
@@ -1,0 +1,1 @@
+Versions/Current/Resources
